:recycle: Refactor check constraint to work with Django 3.2
The implicit version did not work anymore on Django 3.2 as something
in the migration autodetector caused the constraints to get dropped.

Additionally, factoring things out into a contraint by itself is not
straightforward because of the deferred FKs being added _after_ the
virtual field being added to the migration model, making it
impossible to resolve the fields for the check constraint check. The
solution is to defer the constraint evaluation itself until SQL
runtime and only recording the minimal information to combine with the
final model which is passed in as an argument.

Under the hood, we leverage the CheckConstraint logic and instantiate
it just-in-time, so there's no custom SQL involved.

This change will probably cause constraints to be dropped and re-added
on projects upgrading to the new version, but the net result should be
the same.